# apps/derrops-cloud/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cloud-relay/connection": {
      "get": {
        "operationId": "CloudRelayController_findAllConnections",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CloudRelayConnection"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List relay connections for the authenticated user's tenant",
        "tags": [
          "Cloud Relay"
        ]
      },
      "post": {
        "description": "Creates a relay connection. For platform-queue and hybrid delivery modes, an SQS FIFO queue is provisioned automatically (sqs_queue_mode=platform) or the customer-provided queue URL is stored.\n\nThe response includes one-time credentials (IAM access key, Aegis registration token) that are never returned again.\n\ntenantId and userId are read from the verified Cognito id_token — no client-supplied headers are trusted.",
        "operationId": "CloudRelayController_createConnection",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCloudRelayConnectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCloudRelayConnectionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a new relay connection",
        "tags": [
          "Cloud Relay"
        ]
      }
    },
    "/cloud-relay/connection/{id}": {
      "patch": {
        "operationId": "CloudRelayController_updateConnection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Connection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCloudRelayConnectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloudRelayConnection"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a relay connection (name, URL, linked Aegis)",
        "tags": [
          "Cloud Relay"
        ]
      },
      "delete": {
        "operationId": "CloudRelayController_deleteConnection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Connection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a relay connection, its SQS queue, and IAM user (if any)",
        "tags": [
          "Cloud Relay"
        ]
      }
    },
...
    "/cloud-relay/connection/{id}/test-queue": {
      "post": {
        "description": "Sends a canary message to the connection's SQS queue and verifies the send succeeds. Validates that the platform role has sqs:SendMessage permission on the queue.",
        "operationId": "CloudRelayController_testQueueConnection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Connection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sent": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Test SQS queue connectivity for a platform-queue or hybrid relay connection",
        "tags": [
          "Cloud Relay"
        ]
      }
    },
    "/cloud-relay/job": {
      "post": {
        "description": "derrops-cloud routes the request via the connection delivery_mode. direct: result is returned inline (completed immediately). relay-queue/platform-queue: job is pending — poll GET /cloud-relay/job/:id for the result.",
        "operationId": "CloudRelayController_enqueueJob",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCloudRelayJobDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloudRelayJob"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          },
          "503": {
            "description": "Relay unreachable (direct/relay-queue modes)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit a proxy job",
        "tags": [
          "Cloud Relay"
        ]
      }
    },
    "/cloud-relay/job/{id}": {
      "get": {
        "description": "For relay-queue mode, derrops-cloud syncs status from the relay on each poll. For platform-queue and direct modes, returns the stored job state.",
        "operationId": "CloudRelayController_getJob",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job UUID returned by POST /cloud-relay/job",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloudRelayJob"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Poll for the result of a proxy job",
        "tags": [
          "Cloud Relay"
        ]
      }
    },
    "/cloud-relay/queue/next": {
      "get": {
        "description": "The relay polls this endpoint to claim jobs for legacy HTTP-polling connections. SQS-enabled relays (local-dev) receive jobs via SQS long-poll instead. Authenticate with the connection api_key as a Bearer token.",
        "operationId": "CloudRelayController_claimNextJob",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Bearer <connection api_key>",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job claimed — execute and POST result to /cloud-relay/job/:id/result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloudRelayJob"
                }
              }
            }
          },
          "204": {
            "description": "No pending jobs"
          },
          "401": {
            "description": "Invalid api_key"
          }
        },
        "summary": "Claim the next pending platform-queue job (called by the relay)",
        "tags": [
          "Cloud Relay"
        ]
      }
    },
    "/cloud-relay/job/{id}/result": {
      "post": {
        "description": "After executing a claimed job, the relay posts the result here. Set `failed: true` to mark the job as failed.",
        "operationId": "CloudRelayController_deliverJobResult",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Job UUID (from GET /cloud-relay/queue/next or SQS message body)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "Bearer <connection api_key>",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeliverJobResultDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloudRelayJob"
                }
              }
            }
          },
          "401": {
            "description": "Invalid api_key"
          },
          "404": {
            "description": "Job not found"
          }
        },
        "summary": "Deliver the result of a platform-queue job (called by the relay)",
        "tags": [
          "Cloud Relay"
        ]
      }
    }
  },
  "info": {
    "title": "Derrops API",
    "description": "Internal API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CloudRelayConnection": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "5c963787-d89d-4260-adaf-6541c41cb982"
          },
          "tenant_id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "us-east-1 relay"
          },
          "url": {
            "type": "string",
            "example": "https://xyz.execute-api.ap-southeast-2.amazonaws.com/prod",
            "description": "Base URL of the relay instance. Used for direct and relay-queue modes. For platform-queue mode, the relay is not called inbound — leave this as the relay identifier URL."
          },
          "type": {
            "type": "string",
            "enum": [
              "managed",
              "self-hosted",
              "local-dev"
            ],
            "default": "managed",
            "description": "managed     — Derrops-hosted Lambda relay.\nself-hosted — Customer-deployed relay on their own infrastructure.\nlocal-dev   — Developer local machine. delivery_mode is locked to platform-queue. An SQS FIFO queue is provisioned automatically."
          },
          "delivery_mode": {
            "type": "string",
            "enum": [
              "direct",
              "relay-queue",
              "platform-queue",
              "hybrid"
            ],
            "description": "direct         — derrops-cloud calls relay synchronously. Relay must be reachable from derrops-cloud.\nrelay-queue    — derrops-cloud submits to relay queue, polls relay for result. Relay must be reachable from derrops-cloud.\nplatform-queue — relay polls derrops-cloud outbound and posts results back. Use when relay cannot accept inbound connections.\nhybrid         — platform tries direct HTTP first; falls back to platform-queue on failure. Requires both url and sqs_queue_url."
          },
          "sqs_queue_mode": {
            "type": "string",
            "enum": [
              "platform",
              "relay"
            ],
            "description": "platform — Derrops provisions and owns the SQS FIFO queue (default for local-dev). relay    — Customer provisions the queue in their own AWS account and grants the DerropsSqsPublishRole SendMessage access. Use relay mode when the customer's network cannot reach SQS endpoints in the Derrops account."
          },
          "sqs_queue_url": {
            "type": "string",
            "nullable": true,
            "description": "SQS FIFO queue URL for this relay connection. platform mode: provisioned by derrops-cloud and stored here. relay mode: provided by the customer at registration time.",
            "example": "https://sqs.ap-southeast-2.amazonaws.com/123456789/derrops-acme-local-abc123-relay456.fifo"
          },
          "sqs_region": {
            "type": "string",
            "nullable": true,
            "description": "AWS region of the SQS queue.",
            "example": "ap-southeast-2"
          },
          "aegis_id": {
            "type": "string",
            "nullable": true,
            "description": "UUID of the linked AegisInstance. Optional — null means no Aegis for this connection.",
            "example": "9a1b2c3d-4e5f-6789-abcd-ef0123456789"
          },
          "iam_user_arn": {
            "type": "string",
            "nullable": true,
            "description": "ARN of the IAM user provisioned for relay SQS queue access (sqs_queue_mode=platform only). Null until IAM provisioning is implemented."
          },
          "iam_access_key_id": {
            "type": "string",
            "nullable": true,
            "description": "Access key ID of the IAM credential created for the relay. Stored for operator reference only — the secret access key is never stored."
          },
          "api_key": {
            "type": "string",
            "description": "API key for relay authentication. direct/relay-queue: derrops-cloud sends this as Bearer token to the relay. platform-queue: relay sends this as Bearer token when polling derrops-cloud. Configure the relay with RELAY_API_KEY (inbound) or RELAY_PLATFORM_TOKEN (outbound polling)."
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "example": "2026-03-23T10:00:00.000Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "example": "2026-03-23T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "name",
          "url",
          "type",
          "delivery_mode",
          "api_key",
          "created_at",
          "updated_at"
        ]
      },
      "CreateCloudRelayConnectionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "managed",
              "self-hosted",
              "local-dev"
            ],
            "default": "managed",
            "description": "managed     — Derrops-hosted Lambda relay.\nself-hosted — Customer-deployed relay on their own infrastructure.\nlocal-dev   — Developer local machine. delivery_mode is locked to platform-queue and an SQS FIFO queue is provisioned automatically."
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "example": "Production Relay",
            "description": "Human-readable name. Auto-generated if omitted."
          },
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://xyz.execute-api.ap-southeast-2.amazonaws.com/prod",
            "description": "Base URL of the relay instance. Required for direct and relay-queue modes, and for the HTTP path in hybrid mode. Not used for platform-queue connections (relay makes only outbound connections)."
          },
          "delivery_mode": {
            "type": "string",
            "enum": [
              "direct",
              "relay-queue",
              "platform-queue",
              "hybrid"
            ],
            "default": "direct",
            "description": "direct          — derrops-cloud calls relay synchronously.\nrelay-queue     — derrops-cloud submits to relay queue, polls relay for result.\nplatform-queue  — relay polls derrops-cloud outbound. Use when relay cannot accept inbound connections.\nhybrid          — platform tries direct HTTP first, falls back to SQS on failure. Requires both url and sqs_queue_url.\nFor local-dev relays this is always platform-queue regardless of what is sent."
          },
          "sqs_queue_mode": {
            "type": "string",
            "enum": [
              "platform",
              "relay"
            ],
            "default": "platform",
            "description": "platform — Derrops provisions and owns the SQS FIFO queue. relay    — Customer provisions the queue and grants sqs:SendMessage to the Derrops platform role; provide relay_sqs_queue_url. Required when delivery_mode is platform-queue or hybrid."
          },
          "relay_sqs_queue_url": {
            "type": "string",
            "format": "uri",
            "description": "Customer-owned SQS FIFO queue URL. Required when sqs_queue_mode=relay. The customer must grant the DerropsSqsPublishRole sqs:SendMessage permission on this queue.",
            "example": "https://sqs.ap-southeast-2.amazonaws.com/999888777/my-relay-queue.fifo"
          },
          "aegis_id": {
            "type": "string",
            "description": "UUID of an AegisInstance to link to this connection. The Aegis must belong to the same tenant. Null or omitted means no Aegis is linked.",
            "example": "9a1b2c3d-4e5f-6789-abcd-ef0123456789"
          }
        }
      },
      "CreateCloudRelayConnectionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "5c963787-d89d-4260-adaf-6541c41cb982"
          },
          "tenant_id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "us-east-1 relay"
          },
          "url": {
            "type": "string",
            "example": "https://xyz.execute-api.ap-southeast-2.amazonaws.com/prod",
            "description": "Base URL of the relay instance. Used for direct and relay-queue modes. For platform-queue mode, the relay is not called inbound — leave this as the relay identifier URL."
          },
          "type": {
            "type": "string",
            "enum": [
              "managed",
              "self-hosted",
              "local-dev"
            ],
            "default": "managed",
            "description": "managed     — Derrops-hosted Lambda relay.\nself-hosted — Customer-deployed relay on their own infrastructure.\nlocal-dev   — Developer local machine. delivery_mode is locked to platform-queue. An SQS FIFO queue is provisioned automatically."
          },
          "delivery_mode": {
            "type": "string",
            "enum": [
              "direct",
              "relay-queue",
              "platform-queue",
              "hybrid"
            ],
            "description": "direct         — derrops-cloud calls relay synchronously. Relay must be reachable from derrops-cloud.\nrelay-queue    — derrops-cloud submits to relay queue, polls relay for result. Relay must be reachable from derrops-cloud.\nplatform-queue — relay polls derrops-cloud outbound and posts results back. Use when relay cannot accept inbound connections.\nhybrid         — platform tries direct HTTP first; falls back to platform-queue on failure. Requires both url and sqs_queue_url."
          },
          "sqs_queue_mode": {
            "type": "string",
            "enum": [
              "platform",
              "relay"
            ],
            "description": "platform — Derrops provisions and owns the SQS FIFO queue (default for local-dev). relay    — Customer provisions the queue in their own AWS account and grants the DerropsSqsPublishRole SendMessage access. Use relay mode when the customer's network cannot reach SQS endpoints in the Derrops account."
          },
          "sqs_queue_url": {
            "type": "string",
            "nullable": true,
            "description": "SQS FIFO queue URL for this relay connection. platform mode: provisioned by derrops-cloud and stored here. relay mode: provided by the customer at registration time.",
            "example": "https://sqs.ap-southeast-2.amazonaws.com/123456789/derrops-acme-local-abc123-relay456.fifo"
          },
          "sqs_region": {
            "type": "string",
            "nullable": true,
            "description": "AWS region of the SQS queue.",
            "example": "ap-southeast-2"
          },
          "aegis_id": {
            "type": "string",
            "nullable": true,
            "description": "UUID of the linked AegisInstance. Optional — null means no Aegis for this connection.",
            "example": "9a1b2c3d-4e5f-6789-abcd-ef0123456789"
          },
          "iam_user_arn": {
            "type": "string",
            "nullable": true,
            "description": "ARN of the IAM user provisioned for relay SQS queue access (sqs_queue_mode=platform only). Null until IAM provisioning is implemented."
          },
          "iam_access_key_id": {
            "type": "string",
            "nullable": true,
            "description": "Access key ID of the IAM credential created for the relay. Stored for operator reference only — the secret access key is never stored."
          },
          "api_key": {
            "type": "string",
            "description": "API key for relay authentication. direct/relay-queue: derrops-cloud sends this as Bearer token to the relay. platform-queue: relay sends this as Bearer token when polling derrops-cloud. Configure the relay with RELAY_API_KEY (inbound) or RELAY_PLATFORM_TOKEN (outbound polling)."
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "example": "2026-03-23T10:00:00.000Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "example": "2026-03-23T10:00:00.000Z"
          },
          "iam_access_key_id_created": {
            "type": "string",
            "description": "IAM access key ID for the provisioned relay IAM user (sqs_queue_mode=platform only). Returned once — not stored. Null until IAM provisioning is implemented.",
            "example": "AKIAIOSFODNN7EXAMPLE"
          },
          "iam_secret_access_key": {
            "type": "string",
            "description": "IAM secret access key for the provisioned relay IAM user (sqs_queue_mode=platform only). Returned once — never stored. Null until IAM provisioning is implemented.",
            "example": "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
          },
          "aegis_registration_token": {
            "type": "string",
            "description": "One-time Aegis registration token, returned only when a new Aegis instance was registered as part of this connection creation. Never returned again.",
            "example": "eyJhbGciOiJSUzI1NiJ9..."
          }
        },
        "required": [
          "id",
          "tenant_id",
          "name",
          "url",
          "type",
          "delivery_mode",
          "api_key",
          "created_at",
          "updated_at"
        ]
      },
...
      "DeliverJobResultDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}